# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UZV Aquafarm Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pools:
    get:
      operationId: getPools
      tags: [pools]
      summary: Get pools
      parameters:
        - name: farmId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of pools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pool"
    post:
      operationId: createPool
      tags: [pools]
      summary: Create a pool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePoolBody"
      responses:
        "201":
          description: Pool created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pool"
  /pools/{id}:
    get:
      operationId: getPool
      tags: [pools]
      summary: Get a pool by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pool
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pool"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updatePool
      tags: [pools]
      summary: Update a pool
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePoolBody"
      responses:
        "200":
          description: Pool updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pool"
    delete:
      operationId: deletePool
      tags: [pools]
      summary: Delete a pool
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # MEASUREMENTS
...
  /calculators/pool-volume:
    post:
      operationId: calcPoolVolume
      tags: [calculators]
      summary: Calculate pool volume
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PoolVolumeInput"
      responses:
        "200":
          description: Calculation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalculatorResult"
  /calculators/drum-filter:
    post:
      operationId: calcDrumFilter
      tags: [calculators]
      summary: Calculate drum filter performance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DrumFilterInput"
      responses:
        "200":
          description: Calculation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalculatorResult"
  /calculators/pump:
    post:
      operationId: calcPump
      tags: [calculators]
      summary: Calculate pump power
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PumpInput"
      responses:
        "200":
          description: Calculation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalculatorResult"
  /calculators/compressor:
    post:
      operationId: calcCompressor
      tags: [calculators]
      summary: Calculate compressor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompressorInput"
      responses:
        "200":
          description: Calculation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalculatorResult"
  /calculators/oxygen:
    post:
      operationId: calcOxygen
      tags: [calculators]
      summary: Calculate oxygen concentrator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OxygenInput"
      responses:
        "200":
          description: Calculation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalculatorResult"
  /calculators/checklist:
    get:
      operationId: getChecklist
      tags: [calculators]
      summary: Get equipment checklist
      responses:
        "200":
          description: Checklist items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChecklistItem"
    post:
      operationId: saveChecklist
      tags: [calculators]
      summary: Save checklist progress
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/ChecklistItemUpdate"
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # REPORTS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    # AUTH
...
    Pool:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        shape:
          type: string
          enum: [rectangular, circular, oval]
        length:
          type: ["number", "null"]
        width:
          type: ["number", "null"]
        diameter:
          type: ["number", "null"]
        depth:
          type: number
        volume:
          type: number
        farmId:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - name
        - shape
        - depth
        - volume
        - farmId
        - createdAt
    CreatePoolBody:
      type: object
      properties:
        name:
          type: string
        shape:
          type: string
          enum: [rectangular, circular, oval]
        length:
          type: ["number", "null"]
        width:
          type: ["number", "null"]
        diameter:
          type: ["number", "null"]
        depth:
          type: number
        farmId:
          type: integer
      required:
        - name
        - shape
        - depth
        - farmId

    # MEASUREMENTS
...
    PoolVolumeInput:
      type: object
      properties:
        shape:
          type: string
          enum: [rectangular, circular, oval]
        length:
          type: ["number", "null"]
        width:
          type: ["number", "null"]
        diameter:
          type: ["number", "null"]
        depth:
          type: number
      required:
        - shape
        - depth
    DrumFilterInput:
      type: object
      properties:
        poolVolume:
          type: number
        fishBiomass:
          type: number
        feedRate:
          type: number
      required:
        - poolVolume
        - fishBiomass
        - feedRate
    PumpInput:
      type: object
      properties:
        poolVolume:
          type: number
        turnoverRate:
          type: number
        headHeight:
          type: number
      required:
        - poolVolume
        - turnoverRate
        - headHeight
    CompressorInput:
      type: object
      properties:
        poolVolume:
          type: number
        fishBiomass:
          type: number
        waterTemperature:
          type: number
      required:
        - poolVolume
        - fishBiomass
        - waterTemperature
    OxygenInput:
      type: object
      properties:
        poolVolume:
          type: number
        fishBiomass:
          type: number
        targetOxygen:
          type: number
        currentOxygen:
          type: number
      required:
        - poolVolume
        - fishBiomass
        - targetOxygen
        - currentOxygen
    CalculatorResult:
      type: object
      properties:
        result:
          type: number
        unit:
          type: string
        details:
          type: object
          additionalProperties: true
        recommendations:
          type: array
          items:
            type: string
      required:
        - result
        - unit
    ChecklistItem:
      type: object
      properties:
        id:
          type: string
        category:
          type: string
        name:
          type: string
        description:
          type: string
        completed:
          type: boolean
        required:
          type: boolean
      required:
        - id
        - category
        - name
        - description
        - completed
        - required
    ChecklistItemUpdate:
      type: object
      properties:
        id:
          type: string
        completed:
          type: boolean
      required:
        - id
        - completed

    # DASHBOARD
...